Latest digestparser library, pass in the digest_config and set it in the digest_provider.py
