+++ tests/activity/digest.cfg
@@ -20,6 +20,7 @@ output_file_name_pattern: {author}_{msid:0>5}.docx
 iiif_image_uri:
 iiif_image_source_uri:
 iiif_info_url: 
+doi_pattern:
 
 [elife]
 medium_license: cc-40-by
@@ -27,3 +28,4 @@ medium_image_url: https://iiif.elifesciences.org/digests:{msid:0>5}%2F{file_name
 iiif_image_uri: https://iiif.elifesciences.org/digests:{msid:0>5}%2F{file_name}
 iiif_image_source_uri: https://iiif.elifesciences.org/digests:{msid:0>5}%2F{file_name}/full/full/0/default.jpg
 iiif_info_url: https://prod--iiif.elifesciences.org/digests:{msid:0>5}%2F{file_name}/info.json
+doi_pattern: https://doi.org/10.7554/eLife.{msid:0>5}
